Set minimum Node version to 12.x.x
Node 12 has been out for a while now and since this is a standalone tool (not a library that’s used as a dependency of some larger application) there’s no real reason to support old Node versions.

As a bonus, Node 12 has `Array.prototype.flat()` and `Array.prototype.flatMap()` so now we can drop the `core-js` dependency.